:hammer: Move duplicated Container generator code to dedicated util functions
